+++ eth-contracts/contracts/test/GovernanceUpgraded.sol
@@ -4,7 +4,7 @@ import "../Governance.sol";
 
 
 contract GovernanceUpgraded is Governance {
-    function newFunction() public view returns (uint) {
+    function newFunction() public view returns (uint256) {
         _requireIsInitialized();
 
         return 5;
